requirements.txt
certifi==2023.5.7
charset-normalizer==3.1.0
easydict==1.10
filelock==3.12.2
idna==3.4
install==1.3.5
Jinja2==3.1.2
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.1
numpy==1.24.3
opencv-python==4.7.0.72
packaging==23.1
pandas==2.0.2
Pillow==9.5.0
protobuf==3.20.3
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
requests==2.31.0
six==1.16.0
sympy==1.12
tensorboardX==2.6
torch==2.0.1
torchvision==0.15.2
typing_extensions==4.6.3
tzdata==2023.3
urllib3==2.0.3
